refactor: organize confluence tools into tools/ subdirectory
Move individual tool files into tools/ subdirectory for better organization:

Structure:
- packages/tools/src/web/confluence/
  - index.ts (main exports and factory function)
  - types.ts (types and interfaces)
  - auth.ts (auth utilities)
  - tools/ (individual tool implementations)
    - search-confluence.ts
    - get-confluence-page.ts
    - list-confluence-spaces.ts
    - get-space-pages.ts
    - create-confluence-page.ts
    - update-confluence-page.ts
    - archive-confluence-page.ts

Benefits:
- Clearly separates tool implementations from infrastructure
- Similar to web-search/ which has providers/ subdirectory
- Keeps root level clean with just exports, types, and utilities
- Easy to find individual tools

Build passes successfully.
